refactor lifting
