+++ flax/core/__init__.py
@@ -1,3 +1,4 @@
 from .frozen_dict import FrozenDict, freeze, unfreeze
 from .tracers import current_trace, trace_level, check_trace_level
-from .scope import in_kind_filter, Scope, Array, ScanVariableMode, apply, init
+from .scope import in_kind_filter, Scope, Array, apply, init
+from .lift import scan, vmap, jit
